backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
groq==0.11.0